Create SkuidLicenseUse
Skuid Page showing Salesforce Login activity for users that have Skuid Licenses by profile.  This page shows 6 months of login activity. 

If there are more than 100,000 logins a month in your org it will not likely work.  In this case add a condition to the UserLogin model that limits LoginTime to "LAST_MONTH"

Note: This page will not work when a Site License has been implemented by Skuid.